fix: ComputerSetTest & ExecutorTest
